# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MinoDuck MVP API",
    "version": "0.1.0",
    "description": "Internal API behind same-origin BFF. UTC-exclusive end windows, exact decimal strings, no cross-currency sums. Endpoint-specific evidence objects are intentionally extensible."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    },
    {
      "url": "https://api.minoduck.ai"
    }
  ],
  "security": [
    {
      "Session": [],
      "BFFService": []
    }
  ],
  "paths": {
...
    "/api/v1/webhooks/stripe": {
      "post": {
        "operationId": "stripeWebhook",
        "summary": "Stripe webhook",
        "responses": {
          "200": {
            "description": "Successful response; most report/evidence objects are endpoint-specific."
          },
          "default": {
            "description": "Machine-readable error; 402 plan gate, 403 role/CSRF, 404 hidden/unavailable resource, 409 conflict, 503 retryable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "name": "Stripe-Signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/auth/email/start": {
      "post": {
        "operationId": "emailStart",
        "summary": "Email start",
        "responses": {
          "default": {
            "description": "Machine-readable error; 402 plan gate, 403 role/CSRF, 404 hidden/unavailable resource, 409 conflict, 503 retryable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "202": {
            "description": "Successful response; most report/evidence objects are endpoint-specific."
          }
        },
        "security": [
          {
            "BFFService": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailStart"
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/email/verify": {
      "post": {
        "operationId": "emailVerify",
        "summary": "Email verify",
        "responses": {
          "200": {
            "description": "Successful response; most report/evidence objects are endpoint-specific."
          },
          "default": {
            "description": "Machine-readable error; 402 plan gate, 403 role/CSRF, 404 hidden/unavailable resource, 409 conflict, 503 retryable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BFFService": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailVerify"
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/google/start": {
      "get": {
        "operationId": "googleStart",
        "summary": "Google start",
        "responses": {
          "200": {
            "description": "Successful response; most report/evidence objects are endpoint-specific."
          },
          "default": {
            "description": "Machine-readable error; 402 plan gate, 403 role/CSRF, 404 hidden/unavailable resource, 409 conflict, 503 retryable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "302": {
            "description": "Redirect to fixed OAuth or application destination."
          }
        },
        "security": [
          {
            "BFFService": []
          }
        ]
      }
    },
    "/api/v1/auth/google/callback": {
      "get": {
        "operationId": "googleCallback",
        "summary": "Google callback",
        "responses": {
          "200": {
            "description": "Successful response; most report/evidence objects are endpoint-specific."
          },
          "default": {
            "description": "Machine-readable error; 402 plan gate, 403 role/CSRF, 404 hidden/unavailable resource, 409 conflict, 503 retryable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "302": {
            "description": "Redirect to fixed OAuth or application destination."
          }
        },
        "security": [
          {
            "BFFService": []
          }
        ]
      }
    },
    "/api/v1/providers": {
      "get": {
        "operationId": "providers",
        "summary": "Providers",
        "responses": {
          "200": {
            "description": "Successful response; most report/evidence objects are endpoint-specific."
          },
          "default": {
            "description": "Machine-readable error; 402 plan gate, 403 role/CSRF, 404 hidden/unavailable resource, 409 conflict, 503 retryable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BFFService": []
          }
        ]
      }
    },
    "/api/v1/me": {
      "get": {
        "operationId": "me",
        "summary": "Me",
        "responses": {
          "200": {
            "description": "Successful response; most report/evidence objects are endpoint-specific."
          },
          "default": {
            "description": "Machine-readable error; 402 plan gate, 403 role/CSRF, 404 hidden/unavailable resource, 409 conflict, 503 retryable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/preferences": {
      "patch": {
        "operationId": "preferences",
        "summary": "Preferences",
        "responses": {
          "200": {
            "description": "Successful response; most report/evidence objects are endpoint-specific."
          },
          "default": {
            "description": "Machine-readable error; 402 plan gate, 403 role/CSRF, 404 hidden/unavailable resource, 409 conflict, 503 retryable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Preferences"
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout",
        "summary": "Logout",
        "responses": {
          "200": {
            "description": "Successful response; most report/evidence objects are endpoint-specific."
          },
          "default": {
            "description": "Machine-readable error; 402 plan gate, 403 role/CSRF, 404 hidden/unavailable resource, 409 conflict, 503 retryable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/workspaces": {
      "get": {
        "operationId": "listWorkspaces",
        "summary": "List workspaces",
        "responses": {
          "200": {
            "description": "Successful response; most report/evidence objects are endpoint-specific."
          },
          "default": {
            "description": "Machine-readable error; 402 plan gate, 403 role/CSRF, 404 hidden/unavailable resource, 409 conflict, 503 retryable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createWorkspace",
        "summary": "Create workspace",
        "responses": {
          "default": {
            "description": "Machine-readable error; 402 plan gate, 403 role/CSRF, 404 hidden/unavailable resource, 409 conflict, 503 retryable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "201": {
            "description": "Successful response; most report/evidence objects are endpoint-specific."
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspace"
              }
            }
          }
        }
      }
    },
    "/api/v1/invitations/accept": {
      "post": {
        "operationId": "acceptInvitation",
        "summary": "Accept invitation",
        "responses": {
          "200": {
            "description": "Successful response; most report/evidence objects are endpoint-specific."
          },
          "default": {
            "description": "Machine-readable error; 402 plan gate, 403 role/CSRF, 404 hidden/unavailable resource, 409 conflict, 503 retryable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/workspaces/{wid}": {
      "get": {
        "operationId": "workspace",
        "summary": "Workspace",
        "responses": {
          "200": {
            "description": "Successful response; most report/evidence objects are endpoint-specific."
          },
          "default": {
            "description": "Machine-readable error; 402 plan gate, 403 role/CSRF, 404 hidden/unavailable resource, 409 conflict, 503 retryable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "wid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      },
      "patch": {
        "operationId": "updateWorkspace",
        "summary": "Update workspace",
        "responses": {
          "200": {
            "description": "Successful response; most report/evidence objects are endpoint-specific."
          },
          "default": {
            "description": "Machine-readable error; 402 plan gate, 403 role/CSRF, 404 hidden/unavailable resource, 409 conflict, 503 retryable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "wid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/workspaces/{wid}/members": {
      "get": {
        "operationId": "members",
        "summary": "Members",
        "responses": {
          "200": {
            "description": "Successful response; most report/evidence objects are endpoint-specific."
          },
          "default": {
            "description": "Machine-readable error; 402 plan gate, 403 role/CSRF, 404 hidden/unavailable resource, 409 conflict, 503 retryable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "wid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/v1/workspaces/{wid}/invitations": {
      "post": {
        "operationId": "invite",
        "summary": "Invite",
        "responses": {
          "200": {
            "description": "Successful response; most report/evidence objects are endpoint-specific."
          },
          "default": {
            "description": "Machine-readable error; 402 plan gate, 403 role/CSRF, 404 hidden/unavailable resource, 409 conflict, 503 retryable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "wid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/workspaces/{wid}/members/{memberId}": {
      "delete": {
        "operationId": "removeMember",
        "summary": "Remove member",
        "responses": {
          "200": {
            "description": "Successful response; most report/evidence objects are endpoint-specific."
          },
          "default": {
            "description": "Machine-readable error; 402 plan gate, 403 role/CSRF, 404 hidden/unavailable resource, 409 conflict, 503 retryable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "wid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/workspaces/{wid}/deletion-requests": {
      "post": {
        "operationId": "deleteWorkspace",
        "summary": "Delete workspace",
        "responses": {
          "200": {
            "description": "Successful response; most report/evidence objects are endpoint-specific."
          },
          "default": {
            "description": "Machine-readable error; 402 plan gate, 403 role/CSRF, 404 hidden/unavailable resource, 409 conflict, 503 retryable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "wid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/workspaces/{wid}/connections": {
      "get": {
        "operationId": "connections",
        "summary": "Connections",
        "responses": {
          "200": {
            "description": "Successful response; most report/evidence objects are endpoint-specific."
          },
          "default": {
            "description": "Machine-readable error; 402 plan gate, 403 role/CSRF, 404 hidden/unavailable resource, 409 conflict, 503 retryable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "wid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      },
      "post": {
        "operationId": "createConnection",
        "summary": "Create connection",
        "responses": {
          "default": {
            "description": "Machine-readable error; 402 plan gate, 403 role/CSRF, 404 hidden/unavailable resource, 409 conflict, 503 retryable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "201": {
            "description": "Successful response; most report/evidence objects are endpoint-specific."
          }
        },
        "parameters": [
          {
            "name": "wid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConnection"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/workspaces/{wid}/connections/{cid}/sync": {
      "post": {
        "operationId": "syncConnection",
        "summary": "Sync connection",
        "responses": {
          "200": {
            "description": "Successful response; most report/evidence objects are endpoint-specific."
          },
          "default": {
            "description": "Machine-readable error; 402 plan gate, 403 role/CSRF, 404 hidden/unavailable resource, 409 conflict, 503 retryable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "wid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/workspaces/{wid}/insights": {
      "get": {
        "operationId": "insights",
        "summary": "Insights",
        "responses": {
          "200": {
            "description": "Successful response; most report/evidence objects are endpoint-specific."
          },
          "default": {
            "description": "Machine-readable error; 402 plan gate, 403 role/CSRF, 404 hidden/unavailable resource, 409 conflict, 503 retryable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "wid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-md_session",
        "description": "Development uses md_session. Obtain via email or Google login."
      },
      "BFFService": {
        "type": "apiKey",
        "in": "header",
        "name": "X-MinoDuck-Service",
        "description": "Server-only service identity. Browser calls same-origin /api/v1 through Nitro."
      }
    },
    "schemas": {
...
      "APIError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message_key": {
                "type": "string"
              },
              "request_id": {
                "type": "string",
                "format": "uuid"
              },
              "retryable": {
                "type": "boolean"
              }
            },
            "required": ["code"]
          }
        },
        "required": ["error"]
      },
...
      "EmailStart": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "locale": {
            "type": "string",
            "enum": ["en", "zh-hans", "zh-hant"]
          }
        },
        "required": ["email"]
      },
...
    }
  }
}